+++ src/main/resources/db/migration/V3__create_review_table.sql
@@ -3,7 +3,7 @@ use `exam`;
 CREATE TABLE IF NOT EXISTS `reviews`(
     `id` int auto_increment NOT NULL,
     `text` varchar(500) NOT NULL,
-    `mark` int not null ,
+    `mark` double not null ,
     `date` datetime not null ,
     `user_id` int NOT NULL ,
     `place_id` int NOT NULL ,
